requirements.txt
bert-serving-client==1.10.0
bert-serving-server==1.10.0
boto3==1.11.6
botocore==1.14.6
certifi==2019.11.28
chardet==3.0.4
Click==7.0
docutils==0.15.2
GPUtil==1.4.0
idna==2.8
jmespath==0.9.4
joblib==0.14.1
numpy==1.18.1
Pillow==7.0.0
python-dateutil==2.8.1
pyzmq==18.1.1
regex==2020.1.8
requests==2.22.0
s3transfer==0.3.1
sacremoses==0.0.38
sentencepiece==0.1.85
six==1.14.0
termcolor==1.1.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.41.1
transformers==2.3.0
urllib3==1.25.7
